ollama app3/requirements.txt
langchain_core
streamlit
langchain_community
python-dotenv
langserve
fastapi
uvicorn
sse_starlette